requirements.txt
Flask==3.1.0
openai==1.59.8
langchain>=0.3.25,<1.0.0
langchain-community>=0.3.25,<1.0.0
langchain-core>=0.3.76,<1.0.0
chromadb>=1.0.20
python-dotenv==1.0.1
requests==2.32.3
tiktoken==0.8.0
APScheduler==3.11.0

